pyproject.toml
[project]
name = "seminar-bankruptcy-prediction"
version = "1.0.0"
description = "Multi-Dataset Bankruptcy Prediction with Advanced ML Models and Econometric Analysis"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Reebal", email = "student@fh-wedel.de"}
]

dependencies = [
  # Core scientific stack
  "numpy>=2.0.0",
  "pandas>=2.2.0",
  "scipy>=1.16.3",
  # Machine Learning
  "scikit-learn>=1.6.0",
  "imbalanced-learn>=0.14.0",
  "xgboost>=2.1.0",
  "lightgbm>=4.5.0",
  "catboost>=1.2.7",
  # Visualization
  "matplotlib>=3.9.0",
  "seaborn>=0.13.0",
  # Interpretability
  # "shap",
  # Statistics & Econometrics
  "statsmodels>=0.14.0",
  # Data handling
  "pyarrow>=18.0.0",
  "openpyxl>=3.1.0",
  "pyyaml>=6.0.0",
  # Jupyter (for potential interactive work)
  "jupyter>=1.1.0",
  "ipykernel>=6.29.0",
  "shap>=0.49.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bankruptcy_prediction"]

# Interpretability
[project.optional-dependencies]
explain = [
  "shap>=0.46.0"
]
